+++ tests/Integration/ActivationTest.php
@@ -3,10 +3,9 @@
 namespace SnapWP\Helper\Tests\Integration;
 
 use SnapWP\Helper;
-use lucatume\WPBrowser\TestCase\WPTestCase;
+use SnapWP\Helper\Tests\TestCase\IntegrationTestCase;
 
-
-class ActivationTest extends WPTestCase {
+class ActivationTest extends IntegrationTestCase {
 	private $wp_version;
 
 	/**
@@ -22,9 +21,9 @@ class ActivationTest extends WPTestCase {
 	 * Tear down the test.
 	 */
 	public function tearDown(): void {
-		parent::tearDown();
-
 		$GLOBALS['wp_version'] = $this->wp_version;
+
+		parent::tearDown();
 	}
 
 	/**