+++ tests/Integration/DisableIntrospectionRuleTest.php
@@ -7,10 +7,17 @@
 
 namespace SnapWP\Helper\Integration\Tests;
 
+use SnapWP\Helper\Tests\TestCase\IntegrationTestCase;
+
 /**
  * Tests the DisableIntrospectionRule class.
  */
-class DisableIntrospectionRuleTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTestCase {
+class DisableIntrospectionRuleTest extends IntegrationTestCase {
+	/**
+	 * Admin user ID.
+	 *
+	 * @var int
+	 */
 	public $admin;
 
 	/**
@@ -39,11 +46,11 @@ class DisableIntrospectionRuleTest extends \Tests\WPGraphQL\TestCase\WPGraphQLTe
 	 */
 	public function tearDown(): void {
 		// Reset the graphql_general_settings option.
-		$settings = get_option( 'graphql_general_settings' );
-		unset( $settings['public_introspection_enabled'] );
+		$settings                                 = get_option( 'graphql_general_settings' );
+		$settings['public_introspection_enabled'] = 'off';
 		update_option( 'graphql_general_settings', $settings );
 
-		remove_filter( 'graphql_debug_enabled', '__return_false' );
+		remove_all_filters( 'graphql_debug_enabled' );
 
 		wp_delete_user( $this->admin );